+++ tests/test_paste.py
@@ -1,19 +1,20 @@
 import pytest
 
 from yap.models import Paste
+from yap.paste import UNTITLED_FILE
 
 
 def test_create(client, app):
     assert client.get("/").status_code == 200
 
     response = client.post(
-        "/", data=dict(filename="Untitled", contents="some\ncontents", visibility="hidden", expire_in="7 days",)
+        "/", data=dict(filename="foo", contents="some\ncontents", visibility="hidden", expire_in="7 days",)
     )
     assert response.status_code == 302
 
     with app.app_context():
         paste = Paste.query.order_by(Paste.created_at.desc()).first()
-        assert paste.filename == "Untitled"
+        assert paste.filename == "foo"
         assert paste.contents == "some\ncontents"
         assert paste.visibility == Paste.Visibility.hidden
         assert paste.author_ip == "127.0.0.1"
@@ -26,7 +27,6 @@ FILL_REQUIRED_FIELDS = "Please fill all the required fields."
     ("data", "expected_error"),
     [
         (dict(filename="", contents="", visibility="", expire_in=""), FILL_REQUIRED_FIELDS),
-        (dict(filename="", contents="bar", visibility="baz", expire_in="qux"), FILL_REQUIRED_FIELDS),
         (dict(filename="foo", contents="", visibility="baz", expire_in="qux"), FILL_REQUIRED_FIELDS),
         (dict(filename="foo", contents="bar", visibility="", expire_in="qux"), FILL_REQUIRED_FIELDS),
         (dict(filename="foo", contents="bar", visibility="baz", expire_in=""), FILL_REQUIRED_FIELDS),
@@ -43,6 +43,15 @@ FILL_REQUIRED_FIELDS = "Please fill all the required fields."
 def test_create_validation(client, data, expected_error):
     response = client.post("/", data=data)
     assert expected_error in response.get_data(as_text=True)
+
+
+def test_create_with_default_filename(client, app):
+    response = client.post("/", data=dict(filename="", contents="bar", visibility="hidden", expire_in="7 days"))
+    assert response.status_code == 302
+
+    with app.app_context():
+        paste = Paste.query.order_by(Paste.created_at.desc()).first()
+        assert paste.filename == UNTITLED_FILE
 
 
 def test_show(client):