+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # upcoming
 - banner is now a link
+- filename can be now omitted in the form
 
 # 0.1.0
 - basic application (pasting works, lifespan doesn't)